ios mobile 스타일 수정
