+++ src/App.css
@@ -255,6 +255,16 @@ textarea {
   }
 }
 
+input {
+  appearance: none;
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  border-radius: 0;
+  -webkit-border-radius:0;
+  -moz-border-radius:0;
+  outline: 0;
+}
+
 input::-webkit-outer-spin-button,
 input::-webkit-inner-spin-button {
   -webkit-appearance: none;
