+++ src/common/components/Layouts/Footer/index.module.css
@@ -1,5 +1,5 @@
 .container {
-  background-color: #0A0A0A;
+  background-color: #060606;
 
   .wrap {
     width: 100%;
@@ -12,8 +12,8 @@
 
   img.logo {
     width: 148px;
-    height: 40px;
-    object-fit: contain;
+    /* height: 40px; */
+    object-fit: cover;
     display: block;
     margin-bottom: 35px;
   }
