+++ src/pages/About/index.module.css
@@ -7,26 +7,46 @@
     .visual_area {
         text-align: center;
         margin: 0 -16px;
-        padding:86px 16px;
-        background-image: url(images/about/about-visual-banner.png);
+        padding:86px 16px 0;
+        /* background-image: url(images/about/about-visual-banner.png); */
         background-size: cover;
         background-repeat: no-repeat;
         background-position: center;
         position: relative;
         z-index: 5;
+        height: calc(100vh - 138px);
 
+        &::after,
         &::before {
             content: "";
             display: block;
             position: absolute;
-            top: 0;
             left: 0;
-            bottom: 0;
             right: 0;
-            background: linear-gradient(0deg, rgba(255,255,255,0) 93%, rgba(0,0,0,1) 100%);
+            bottom: 0;
+            background: linear-gradient(180deg, rgba(255,255,255,0) 93%, #060606 100%);
             z-index: -1;
+            height: 100%;
         }
-        
+
+        &::before {
+            bottom: initial;
+            top: 0;
+            background: linear-gradient(0deg, rgba(255,255,255,0) 90%, #040404 100%);
+            z-index: 1;
+        }
+
+    }
+
+    .video_track {
+        width: 100vw;
+        height: 100%;
+        position: absolute;
+        top: 0;
+        left: 50%;
+        transform: translateX(-50vw);
+        z-index: -1;
+        object-fit: cover;
     }
     
     .title {
@@ -60,6 +80,7 @@
         font-size: 4.267vw;
         font-weight: 400;
         line-height: 6.133vw; 
+        display: none;
     }
 
     .feature_area {
@@ -111,7 +132,7 @@
             font-size: 4.267vw;
             font-weight: 400;
             line-height: 6.133vw;
-            white-space: pre-wrap;
+            white-space: initial;
             color: #A4A4A4;
         }
     }
@@ -146,7 +167,8 @@
         padding: 0 16px;
 
         .visual_area {
-            padding: 200px 16px 352px;
+            padding: 200px 16px 0;
+            height: calc(100vh - 200px);
 
             &::before {
                 background: linear-gradient(0deg, rgba(255,255,255,0) 87%, rgba(0,0,0,1) 100%);
@@ -178,7 +200,7 @@
         }
 
         .feature_area {
-            padding: 270px 0 260px;
+            padding: 170px 0 260px;
 
             .items {
                 grid-template-columns: 1fr 0.7fr;
@@ -228,6 +250,7 @@
                 color: #9A9A9A;
                 line-height: normal;
                 line-height: 45px;
+                white-space: pre-wrap;
             }
         }
