+++ src/pages/Contact/index.module.css
@@ -22,6 +22,7 @@
 
     label {
       position: relative;
+      display: block;
     }
 
     input,
